use action method dynamically with parameters
