+++ frontend/src/pages/EventEdit.js
@@ -5,7 +5,7 @@ import EventForm from "../components/EventForm"
 const EventEditPage = () => {
   const loadedData = useRouteLoaderData("event-detail");
 
-  return <EventForm event={loadedData.event} />
+  return <EventForm method="patch" event={loadedData.event} />
 };
 
 export default EventEditPage;
